Rework the address fluent interface for addresses.
In C++ we may design some parts of design to be provided by implementation,
not interface, initially.  In Rust all the details must be part of the interface
or else generic functions can't use it.  We had no need for generic implementation
for addresses thus it's probably Ok to provide it once.