Add 𝔠𝔰𝔯 registers from Volume II: Privileged Architecture.
Note: while most of these registers are privileges, some of them are not.
We may also want to add faster validity checking later: currently we are just
using naive calculations but it's possible that using 512 byte array would be
faster (or maybe we need to separate userspace-only assembler from full one).