Add preliminary implementation of 𝒎𝒂𝒄𝒉𝒊𝒏𝒆_𝒄𝒐𝒅𝒆 traits.
The idea is to collect code in blocks which are then attached to the main block
and emitted with recalculated expressions related to labels.

This architecture allows easy support for constant pools (they can be added to
separate blocks and then placed in the suitable places), can easily support
𝔯𝔦𝔭-addressing (label can be placed after instruction and distance between two
labels becomes proper 𝖽𝗂𝗌𝗉) or RISC-V 𝗉𝖼 addressing (label is stored before
instruction and expressions of manual are used to calculate two immediates) but
requires post-processing at the time where code is actually emitted.

The hope is that this post-processing wouldn't be too expensive and wouldn't
require simpler and faster assembler for the direct-code-emit phase.

To facilitate that we are trying to not allocate memory in our label
expressions unless they become too complicated to fit in fixed-size buffer.