Return “compressed instruction type” from slow decoder.
And ensure that “compressed instruction type” is the same for all primary
opcodes (we manually merge few “compressed instruction types” and split some
others to make it happen.

Instructions like 𝔠_𝔞𝔡𝔡𝔦16𝔰𝔭 are the most complicated ones: when we do primary
decoding we don't yet know whether we are dealing with 𝔠_𝔞𝔡𝔡𝔦16𝔰𝔭 or 𝔠_𝔩𝔲𝔦
but these instructions have very different ways to handle immediate!

We probably would need to do further adjustments to numbering later
(e.g. we would probably need to add 𝔠_𝔞𝔡𝔡𝔦16𝔰𝔭 before 𝔠_𝔩𝔲𝔦_𝔭𝔬𝔰𝔦𝔱𝔦𝔳𝔢_00),
but this would be done in follow-up CLs.