Implement compressed instructions decoder.
Also make 𝒊𝒏𝒔𝒕𝒓𝒖𝒄𝒕𝒊𝒐𝒏𝒔_𝒄𝒐𝒏𝒔𝒖𝒎𝒆𝒓 traits accept Self instead of &mut Self.

This is easier for testing but means that both parcels producer and also
instructions consumer can only be passed into decode function once.

But if that's not desirable then it's easy to implement appropriate traits
for references thus this approach is strictly more flexible.