Make 𝐧𝐚𝐭𝐢𝐯𝐞_𝐦𝐚𝐜𝐡𝐢𝐧𝐞_𝐜𝐨𝐝𝐞 generic (supports i16/i32/i64/isize).
While isize displacement is enough for native compilation we may also want
to generate code for a foreign architecture and this is realitively simple
change.

The only complication is https://github.com/rust-lang/rust/issues/20671
Because of that limitation we have a lot of code duplication in that module.

Also add smoke tests to verify that all supported operations work.