Fix miri's complaints about pointer without provenance.
We are converting `usize` integer into pointer-to-function. This is grey area
operation which works correctly on most CPUs but may not work in all cases and
is currently not supported by Miri.

Put actual function pointers into the map and use them instead of transmuting.
This still leaves the question of whether that transmute is valid open, but
allows us to verify other parts of code.